requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pandas==2.2.2
openpyxl==3.1.2
psycopg2-binary==2.9.9
APScheduler==3.10.4
paramiko==3.2.0
pytz==2025.2
